clean cache
Hye!! Today i want to share to all about how to clean cache/temp which is one of cause be your computer slow to process. this cache also 'eat' your memory computer so, it must be clear it to make your computer smooth.... This method i learned with my pro friend. it success to make my computer process fast than before it. Read this method till last word before do-it. 

1. Click START and type "run" at the search windows and press Enter. The tab "Run" will show like the picture in this file. 
2. Then type the keyword as the picture's file on the Run. e.g: "recent" and press Enter again. 
3. After that, select all the document in the that window and click delete. 
4. Then, go to the next keyword, then do-it again ways  'number 3'  so on till last keyword except the keyword "tree" ...

** The keyword "tree" for trace all files in your computer . When you type the keyword "tree" first before clean cache , it will take along time to finish tracing than type the keyword "tree"  again at the last